requirements.txt
Pillow==9.3.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
